Password change minor changes
Removed code from the menu.tagx regarding Setting button. I want it the
have a drop down menu.
Added Setting button in the nevbar and set up permissions that it is
only visible for the First Responder.
Added label Password change to the menu Settings.
Added label Password change in the User menu and set permissions that
is only visible for the Admin.